+++ www/go/core/Language.php
@@ -44,16 +44,9 @@ class Language {
 	 */
 	public function getIsoCode(): string
 	{
-
 		if(isset($_GET['SET_LANGUAGE']) && $this->hasLanguage($_GET['SET_LANGUAGE'])) {
 			if(!headers_sent()) {
-				//setcookie('GO_LANGUAGE', $_GET['SET_LANGUAGE'], time() + (10 * 365 * 24 * 60 * 60));
-				Response::get()->setCookie('GO_LANGUAGE', $_GET['SET_LANGUAGE'], [
-					'expires' => time() + (10 * 365 * 24 * 60 * 60),
-					"path" => "/",
-					"samesite" => "Strict",
-					"domain" => Request::get()->getHost()
-				]);
+				setcookie("GO_LANGUAGE", $_GET['SET_LANGUAGE'],time() + (10 * 365 * 24 * 60 * 60),"/",Request::get()->getHost());
 			}
 			return $_GET['SET_LANGUAGE'];
 		}
@@ -64,12 +57,7 @@ class Language {
 			} else {
 				$this->isoCode = go()->getAuthState() && go()->getAuthState()->isAuthenticated() ? go()->getAuthState()->getUser(['language'])->language : $this->getBrowserLanguage();
 				if(!headers_sent()) {
-					Response::get()->setCookie('GO_LANGUAGE', $this->isoCode, [
-						'expires' => time() + (10 * 365 * 24 * 60 * 60),
-						"path" => "/",
-						"samesite" => "Strict",
-						"domain" => Request::get()->getHost()
-					]);
+					setcookie("GO_LANGUAGE", $this->isoCode,time() + (10 * 365 * 24 * 60 * 60),"/",Request::get()->getHost());
 				}
 			}
 		}